+++ src/main/java/com/programmers/domain/post/ui/PostController.java
@@ -1,6 +1,6 @@
 package com.programmers.domain.post.ui;
 
-import com.programmers.domain.post.application.PostService;
+import com.programmers.domain.post.application.PostServiceImpl;
 import com.programmers.domain.post.ui.dto.PostDto;
 import com.programmers.domain.post.ui.dto.PostUpdateDto;
 import lombok.RequiredArgsConstructor;
@@ -16,26 +16,26 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/api")
 public class PostController {
 
-    private final PostService postService;
+    private final PostServiceImpl postServiceImpl;
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping("/posts")
     public Long createPost(@RequestBody PostDto postDto) {
-        return postService.createPost(postDto);
+        return postServiceImpl.createPost(postDto);
     }
 
     @GetMapping("/posts")
     public Page<PostDto> findPostList(@PageableDefault(size = 20, sort = "id", direction = Sort.Direction.DESC) Pageable pageable) {
-        return postService.findAll(pageable);
+        return postServiceImpl.findAll(pageable);
     }
 
     @GetMapping("/posts/{id}")
     public PostDto findPost(@PathVariable Long id) {
-        return postService.findPost(id);
+        return postServiceImpl.findPost(id);
     }
 
     @PatchMapping("/posts/{id}")
     public PostDto updatePost(@PathVariable Long id, @RequestBody PostUpdateDto postUpdateDto) {
-        return postService.updatePost(postUpdateDto, id);
+        return postServiceImpl.updatePost(postUpdateDto, id);
     }
 }
